first attempt at all one file
